storage: re-organise crate into `arena` and `stores` modules
