--- compiler/hash-storage/src/collections/mod.rs
@@ -1,5 +1,0 @@
-//! Various collection types that store data inside the memory arena provided by
-//! [`Wall`](super::Wall).
-
-pub mod row;
-pub mod string;
